requirements.txt
fastapi==0.85.0
requests==2.28.1
uvicorn==0.18.3
pydantic~=1.10.2
starlette~=0.20.4
aiosqlite~=0.17.0
passlib~=1.7.4
python-jose
bcrypt
python-multipart